% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biothings-package.R
\docType{package}
\name{biothings-package}
\alias{biothings}
\alias{biothings-package}
\title{A package for accessing BioThings APIs with R.}
\description{
\href{http://biothings.io/}{BioThings} APIs:
\itemize{
  \item \href{http://mygene.info/}{MyGene}
  \item \href{http://myvariant.info/}{MyVariant}
  \item \href{http://t.biothings.io/}{MyTaxon}
  \item \href{http://mychem.info/}{MyChem}
  ...
}
}
\author{
\strong{Maintainer}: Thomas Johnson \email{thomascjohnson@gmail.com}

}
