% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biothings_clients.R
\docType{data}
\name{biothings_clients}
\alias{biothings_clients}
\title{BioThings API Client Configuration}
\format{An object of class \code{list} of length 4.}
\usage{
biothings_clients
}
\description{
This is a list that containts configuration information for each Biothings
API:
}
\details{
\href{http://biothings.io/}{BioThings} APIs:
\itemize{
  \item \href{http://mygene.info/}{MyGene}
  \item \href{http://myvariant.info/}{MyVariant}
  \item \href{http://t.biothings.io/}{MyTaxon}
  \item \href{http://mychem.info/}{MyChem}
}

To integrate a new API into the package, be sure to update this in addition
to generating the associated methods from getThing and getThings.
}
\keyword{datasets}
