% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioThingsClient.R
\docType{class}
\name{BioThingsClient-class}
\alias{BioThingsClient-class}
\alias{BioThingsClient}
\title{BioThingsClient}
\value{
An S4 class object of the class Biothings.
}
\description{
An S4 Class to access BioThings APIs.

\href{http://biothings.io/}{BioThings} APIs:
\itemize{
  \item \href{http://mygene.info/}{MyGene}
  \item \href{http://myvariant.info/}{MyVariant}
  \item \href{http://t.biothings.io/}{MyTaxon}
  \item \href{http://mychem.info/}{MyChem}
  ...
}
}
\section{Slots}{

\describe{
\item{\code{client}}{A client configuration list.}

\item{\code{version}}{The version of the BioThings package.}

\item{\code{verbose}}{logical.}

\item{\code{debug}}{logical.}
}}

\examples{
biothings <- BioThingsClient("gene")
slot(biothings, "verbose") <- FALSE # default is TRUE
biothings
}
