% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{methods}
\name{btFields}
\alias{btFields}
\alias{btFields,BioThingsClient-method}
\alias{btFields,BioThings,BioThings-method}
\alias{btFields,character-method}
\alias{btFields,BioThings,BioThings-method}
\title{btFields}
\usage{
btFields(biothings, ...)

\S4method{btFields}{BioThingsClient}(biothings, ...)

\S4method{btFields}{character}(biothings, ...)
}
\arguments{
\item{biothings}{A BioThings class object}

\item{...}{Additional API parameters. See API documentation.}
}
\value{
A list of results from the metadata fields endpoint
}
\description{
Get field metadata information from BioThings API field metadata endpoints
}
\examples{
fields <- btFields("gene")
head(fields)
}
