% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\docType{methods}
\name{btGet}
\alias{btGet}
\alias{btGet,BioThings-method}
\alias{btGet,character-method}
\alias{btGet,missing-method}
\title{Get items from BioThings APIs}
\usage{
btGet(biothings, ids, fields, ..., return.as = "records")

\S4method{btGet}{BioThings}(biothings, ids, fields, ...,
  return.as = "records")

\S4method{btGet}{character}(biothings, ids, fields, ...,
  return.as = "records")

\S4method{btGet}{missing}(biothings, ids, fields, ..., return.as = "records")
}
\arguments{
\item{biothings}{BioThings object or character .}

\item{ids}{character.}

\item{fields}{Specify fields to receive from API}

\item{...}{Other parameters to pass to API}

\item{return.as}{character.}
}
\value{
The API response in the form of the provided argument return.as
}
\description{
Use gene, taxon, variant, chemical and other IDs to get annotation
information from BioThings APIs.
}
\examples{
#Equivalent:

btGet("gene", "1017", fields = c("symbol","name","taxid","entrezgene"),
      return.as = "text")
}
