% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{methods}
\name{btMetadata}
\alias{btMetadata}
\alias{btMetadata,BioThingsClient-method}
\alias{btMetadata,BioThings,BioThings-method}
\alias{btMetadata,character-method}
\alias{btMetadata,BioThings,BioThings-method}
\title{btMetadata}
\usage{
btMetadata(biothings, ...)

\S4method{btMetadata}{BioThingsClient}(biothings, ...)

\S4method{btMetadata}{character}(biothings, ...)
}
\arguments{
\item{biothings}{A BioThings class object}

\item{...}{Additional API parameters. See API documentation.}
}
\value{
A list of results from the metadata endpoint
}
\description{
Get metadata information from BioThings API metadata endpoints
}
\examples{
metadata <- btMetadata("gene")
head(metadata)
}
