% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{methods}
\name{getFields}
\alias{getFields}
\alias{getFields,BioThings-method}
\alias{getFields,BioThings,BioThings-method}
\alias{getFields,missing-method}
\alias{getFields,BioThings,BioThings-method}
\title{getFields}
\usage{
getFields(client, ..., biothings)

\S4method{getFields}{BioThings}(client, ..., biothings)

\S4method{getFields}{missing}(client, ..., biothings)
}
\arguments{
\item{client}{A biothings client name: gene, variant, taxon or chem}

\item{...}{Additional API parameters. See API documentation.}

\item{biothings}{A BioThings class object}
}
\value{
A list of results from the metadata fields endpoint
}
\description{
Get field metadata information from BioThings API field metadata endpoints
}
\examples{
fields <- getFields("gene")
head(fields)
}
