% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{methods}
\name{getMetadata}
\alias{getMetadata}
\alias{getMetadata,BioThings-method}
\alias{getMetadata,BioThings,BioThings-method}
\alias{getMetadata,missing-method}
\alias{getMetadata,BioThings,BioThings-method}
\title{getMetadata}
\usage{
getMetadata(client, ..., biothings)

\S4method{getMetadata}{BioThings}(client, ..., biothings)

\S4method{getMetadata}{missing}(client, ..., biothings)
}
\arguments{
\item{client}{A biothings client name: gene, variant, taxon or chem}

\item{...}{Additional API parameters. See API documentation.}

\item{biothings}{A BioThings class object}
}
\value{
A list of results from the metadata endpoint
}
\description{
Get metadata information from BioThings API metadata endpoints
}
\examples{
metadata <- getMetadata("gene")
head(metadata)
}
