% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R, R/Chemical.R, R/Gene.R, R/Taxon.R,
%   R/Variant.R
\docType{methods}
\name{getThing}
\alias{getThing}
\alias{getThing,BioThings-method}
\alias{getThing,missing-method}
\alias{getThings}
\alias{getThings,BioThings-method}
\alias{getThings,missing-method}
\alias{getChemical}
\alias{getChem}
\alias{getChemicals}
\alias{getChems}
\alias{getGene}
\alias{getGene,BioThings-method}
\alias{getGene,missing-method}
\alias{getGenes}
\alias{getGenes,BioThings-method}
\alias{getGenes,missing-method}
\alias{getTaxon}
\alias{getTaxons}
\alias{getVariant}
\alias{getVariants}
\title{Get items from BioThings APIs}
\usage{
getThing(thing, client, fields, ..., return.as, biothings)

\S4method{getThing}{BioThings}(thing, client, fields, ..., return.as, biothings)

\S4method{getThing}{missing}(thing, client, fields, ..., return.as, biothings)

getThings(things, client, fields, ..., return.as, biothings)

\S4method{getThings}{BioThings}(things, client, fields, ..., return.as,
  biothings)

\S4method{getThings}{missing}(things, client, fields, ..., return.as, biothings)

getChemical(chemid, fields = NULL, ..., return.as = c("records", "text"),
  biothings)

getChemical(chemid, fields = NULL, ..., return.as = c("records", "text"),
  biothings)

getChemicals(chemids, fields = NULL, ..., return.as = c("records", "text",
  "data.frame"), biothings)

getChemicals(chemids, fields = NULL, ..., return.as = c("records", "text",
  "data.frame"), biothings)

getGene(geneid, fields = c("symbol", "name", "taxid", "entrezgene"), ...,
  return.as = c("records", "text", "data.frame"), biothings)

\S4method{getGene}{BioThings}(geneid, fields = c("symbol", "name", "taxid",
  "entrezgene"), ..., return.as = c("records", "text", "data.frame"),
  biothings)

\S4method{getGene}{missing}(geneid, fields = c("symbol", "name", "taxid",
  "entrezgene"), ..., return.as = c("records", "text", "data.frame"),
  biothings)

getGenes(geneids, fields = c("symbol", "name", "taxid", "entrezgene"), ...,
  return.as = c("records", "text", "data.frame"), biothings)

\S4method{getGenes}{BioThings}(geneids, fields = c("symbol", "name", "taxid",
  "entrezgene"), ..., return.as = c("records", "text", "data.frame"),
  biothings)

\S4method{getGenes}{missing}(geneids, fields = c("symbol", "name", "taxid",
  "entrezgene"), ..., return.as = c("records", "text", "data.frame"),
  biothings)

getTaxon(taxonid, fields = NULL, ..., return.as = c("records", "text",
  "data.frame"), biothings)

getTaxons(taxonids, fields = NULL, ..., return.as = c("records", "text",
  "data.frame"), biothings)

getVariant(variantid, fields = NULL, ..., return.as = c("records", "text",
  "data.frame"), biothings)

getVariants(variantids, fields = NULL, ..., return.as = c("records", "text",
  "data.frame"), biothings)
}
\arguments{
\item{thing}{character.}

\item{client}{A BioThings client name}

\item{fields}{Specify fields to receive from API}

\item{...}{Other parameters to pass to API}

\item{return.as}{character.}

\item{biothings}{BioThings.}

\item{things}{vector.}

\item{chemid}{character.}

\item{chemids}{vector.}

\item{geneid}{character.}

\item{geneids}{vector.}

\item{taxonid}{character.}

\item{taxonids}{vector.}

\item{variantid}{character.}

\item{variantids}{vector.}
}
\value{
The API response in the form of the provided argument return.as
}
\description{
Use gene, taxon, variant, chemical and other IDs to get annotation information from BioThings APIs.
}
\examples{
#Equivalent:
getThing("1017", "gene", fields = c("symbol","name","taxid","entrezgene"),
          return.as = "text")
getGene("1017", return.as = "text")

# Also equivalent
getThings(c("1017","1018","ENSG00000148795"), "gene",
          fields = c("symbol","name","taxid","entrezgene"),
          return.as = "text")
getGenes(c("1017","1018","ENSG00000148795"), return.as = "text")
}
