% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\alias{queryMany}
\title{query
Retrieve results from the query endpoint of BioThings APIs}
\usage{
query(q, client, ..., return.as = "data.frame", biothings)

queryMany(qterms, client, scopes = NULL, ..., return.as = "data.frame",
  biothings)
}
\arguments{
\item{q}{A query string}

\item{client}{A BioThings client name}

\item{...}{Any parameters to pass to API}

\item{return.as}{Type of return value}

\item{biothings}{An S4 class BioThings object}

\item{qterms}{A vector of query strings}
}
\value{
Returns the API result as the provided return.as type
}
\examples{
query(q="NM_013993", client = "gene")
}
