% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioThings.R
\docType{class}
\name{BioThings-class}
\alias{BioThings-class}
\alias{BioThings}
\title{BioThings - An S4 Class to access BioThings APIs.}
\value{
An S4 class object of the class Biothings.
}
\description{
BioThings - An S4 Class to access BioThings APIs.
}
\section{Slots}{

\describe{
\item{\code{clients}}{A nested list of BioThings API configurations.}

\item{\code{version}}{The version of the BioThings package.}

\item{\code{verbose}}{logical.}

\item{\code{debug}}{logical.}
}}

\examples{
biothings <- new("BioThings")
biothings@verbose <- FALSE # default is TRUE
biothings
}
