% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{methods}
\name{btFields}
\alias{btFields}
\alias{btFields,BioThings-method}
\alias{btFields,BioThings,BioThings-method}
\alias{btFields,missing-method}
\alias{btFields,BioThings,BioThings-method}
\title{btFields}
\usage{
btFields(client, ..., biothings)

\S4method{btFields}{BioThings}(client, ..., biothings)

\S4method{btFields}{missing}(client, ..., biothings)
}
\arguments{
\item{client}{A biothings client name: gene, variant, taxon or chem}

\item{...}{Additional API parameters. See API documentation.}

\item{biothings}{A BioThings class object}
}
\value{
A list of results from the metadata fields endpoint
}
\description{
Get field metadata information from BioThings API field metadata endpoints
}
\examples{
fields <- btFields("gene")
head(fields)
}
