% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R, R/Chemical.R, R/Gene.R, R/Taxon.R,
%   R/Variant.R
\name{getThing}
\alias{getThing}
\alias{getThings}
\alias{getChemical}
\alias{getChem}
\alias{getChemicals}
\alias{getChems}
\alias{getGene}
\alias{getGenes}
\alias{getTaxon}
\alias{getTaxons}
\alias{getVariant}
\alias{getVariants}
\title{Get items from BioThings APIs}
\usage{
getThing(thing, client, fields, ..., return.as, biothings)

getThings(things, client, fields, ..., return.as, biothings)

getChemical(chemid, fields = NULL, ..., return.as = c("records", "text"),
  biothings)

getChemical(chemid, fields = NULL, ..., return.as = c("records", "text"),
  biothings)

getChemicals(chemids, fields = NULL, ..., return.as = c("records", "text",
  "data.frame"), biothings)

getChemicals(chemids, fields = NULL, ..., return.as = c("records", "text",
  "data.frame"), biothings)

getGene(geneid, fields = c("symbol", "name", "taxid", "entrezgene"), ...,
  return.as = c("records", "text", "data.frame"), biothings)

getGenes(geneids, fields = c("symbol", "name", "taxid", "entrezgene"), ...,
  return.as = c("records", "text", "data.frame"), biothings)

getTaxon(taxonid, fields = NULL, ..., return.as = c("records", "text",
  "data.frame"), biothings)

getTaxons(taxonids, fields = NULL, ..., return.as = c("records", "text",
  "data.frame"), biothings)

getVariant(variantid, fields = NULL, ..., return.as = c("records", "text",
  "data.frame"), biothings)

getVariants(variantids, fields = NULL, ..., return.as = c("records", "text",
  "data.frame"), biothings)
}
\arguments{
\item{thing}{character.}

\item{client}{character.}

\item{fields}{vector.}

\item{return.as}{character.}

\item{biothings}{BioThings.}

\item{things}{vector.}
}
\value{
The API response in the form of the provided argument return.as
}
\examples{
#Equivalent:
getThing("1017", "gene", fields = c("symbol","name","taxid","entrezgene"),
          return.as = "text")
getGene("1017", return.as = "text")

# Also equivalent
getThings(c("1017","1018","ENSG00000148795"), "gene",
          fields = c("symbol","name","taxid","entrezgene"),
          return.as = "text")
getGenes(c("1017","1018","ENSG00000148795"), return.as = "text")
}
